requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
defusedxml==0.7.1
Deprecated==1.3.1
deprecation==2.1.0
dotenv==0.9.9
easyocr==1.7.2
fastapi==0.128.0
filelock==3.20.3
fonttools==4.61.1
fpdf2==2.8.5
fsspec==2026.1.0
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-documentai==3.9.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
idna==3.11
ImageIO==2.37.2
img2pdf==0.6.3
Jinja2==3.1.6
joblib==1.5.3
lazy_loader==0.4

lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
nltk==3.9.2
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
ocrmypdf==17.1.0
opencv-python==4.13.0.90
opencv-python-headless==4.13.0.90
packaging==26.0
passlib==1.7.4
pdf2image==1.17.0
pdfminer.six==20260107
pi_heif==1.2.0
pikepdf==10.3.0
pillow==12.1.0
pluggy==1.6.0
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.10
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-jose==3.3.0
bcrypt==4.2.1
pypdfium2==5.3.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2026.1.15
requests==2.32.5
httpx>=0.28.1
rich==14.3.2
rsa==4.9.1
scikit-image==0.26.0
scipy==1.17.0
setuptools==80.10.2
shapely==2.1.2
SQLAlchemy==2.0.36
starlette==0.50.0
sympy==1.14.0
textblob==0.19.0
tifffile==2026.1.28
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.2
triton==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uharfbuzz==0.53.3
urllib3==2.6.3
uvicorn==0.40.0
wrapt==2.1.1
